#2 - fix : 돈 관련 필드 자료형 수정
돈 관련 필드 자료형 BigDecimal 통일

+++ generator-core/src/main/java/com/simpaylog/generatorcore/entity/OccupationalWage.java
@@ -5,6 +5,7 @@ import lombok.AllArgsConstructor;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 import lombok.Setter;
+import java.math.BigDecimal;
 
 @Entity
 @Table(name = "occupational_wages")
@@ -19,5 +20,5 @@ public class OccupationalWage{
 
     private String occupation;
 
-    private Integer monthlyWage;
+    private BigDecimal monthlyWage;
 }
